feat: move integration tests to subfolder
