+++ tests/integration/test_process.py
@@ -3,6 +3,7 @@ import os
 import pickle
 
 import pandas as pd
+import pytest
 
 from importer.configuration.config import Config
 from importer.metadata import Metadata
@@ -31,37 +32,39 @@ def test_create_options(cfg: Config):
     assert opts.ignore_columns is not None
     assert opts.db_uri().startswith("postgresql+psycopg://")
 
+@pytest.mark.integration
+@pytest.mark.skip(reason="Requires live database connection")
+class TestImportService:
+    def test_import_reduced_submission(self, cfg: Config):
+        target_filename: str = "data/output/building_dendro_reduced.xml"
+        expected_filename: str = "tests/test_data/building_dendro_reduced.xml"
 
-def test_import_reduced_submission(cfg: Config):
-    target_filename: str = "data/output/building_dendro_reduced.xml"
-    expected_filename: str = "tests/test_data/building_dendro_reduced.xml"
+        opts: Options = Options(
+            **{
+                "filename": "tests/test_data/building_dendro_reduced.xlsx",
+                "data_types": "dendrochronology",
+                "database": cfg.get("options:database"),
+                "output_folder": "data/output",
+                "skip": False,
+                "submission_id": None,
+                "table_names": None,
+                "xml_filename": None,
+                "check_only": False,
+                "register": False,
+                "explode": False,
+                "timestamp": False,
+                "tidy_xml": False,
+            }
+        )
 
-    opts: Options = Options(
-        **{
-            "filename": "tests/test_data/building_dendro_reduced.xlsx",
-            "data_types": "dendrochronology",
-            "database": cfg.get("options:database"),
-            "output_folder": "data/output",
-            "skip": False,
-            "submission_id": None,
-            "table_names": None,
-            "xml_filename": None,
-            "check_only": False,
-            "register": False,
-            "explode": False,
-            "timestamp": False,
-            "tidy_xml": False,
-        }
-    )
+        metadata: Metadata = Metadata(opts.db_uri())
 
-    metadata: Metadata = Metadata(opts.db_uri())
+        submission: Submission = Submission.load(metadata=metadata, source=opts.filename)
 
-    submission: Submission = Submission.load(metadata=metadata, source=opts.filename)
-
-    service: ImportService = ImportService(metadata=metadata, opts=opts)
-    service.process(submission=submission)
-    assert len(service.specification.errors) == 0
-    assert filecmp.cmp(target_filename, expected_filename, shallow=False)
+        service: ImportService = ImportService(metadata=metadata, opts=opts)
+        service.process(submission=submission)
+        assert len(service.specification.errors) == 0
+        assert filecmp.cmp(target_filename, expected_filename, shallow=False)
 
 
 def load_or_cache_submission(opts, metadata) -> Submission:
